+++ packages/polaris-viz/src/components/LineChart/components/GradientArea/GradientArea.tsx
@@ -1,13 +1,13 @@
 import React, {useMemo} from 'react';
 import type {ScaleLinear} from 'd3-scale';
 import {area} from 'd3-shape';
+import type {DataPoint} from '@shopify/polaris-viz-core';
 import {
   LinearGradientWithStops,
   curveStepRounded,
   uniqueId,
 } from '@shopify/polaris-viz-core';
 
-import type {DataPoint} from '../../../../types';
 import type {DataWithDefaults} from '../../types';
 
 import {getGradientDetails} from './utilities/get-gradient-details';