+++ packages/polaris-viz/src/components/SimpleBarChart/Chart.tsx
@@ -1,6 +1,10 @@
 import React, {useCallback, useMemo} from 'react';
 import {uniqueId} from '@shopify/polaris-viz-core';
-import type {ChartType, DataSeries} from '@shopify/polaris-viz-core';
+import type {
+  ChartType,
+  DataSeries,
+  Dimensions,
+} from '@shopify/polaris-viz-core';
 
 import {LegendContainer, useLegend} from '../../components/LegendContainer';
 import {GradientDefs, HorizontalGroup} from '../shared';
@@ -19,7 +23,6 @@ import {
   HORIZONTAL_BAR_GROUP_DELAY,
   COLOR_VISION_SINGLE_ITEM,
 } from '../../constants';
-import type {Dimensions} from '../../types';
 
 import type {XAxisOptions} from './types';
 import styles from './Chart.scss';