+++ packages/polaris-viz/src/components/TooltipWrapper/TooltipWrapper.tsx
@@ -6,9 +6,9 @@ import React, {
   useMemo,
   useCallback,
 } from 'react';
-import type {DataType} from '@shopify/polaris-viz-core';
+import type {DataType, Dimensions} from '@shopify/polaris-viz-core';
 
-import type {Dimensions, Margin} from '../../types';
+import type {Margin} from '../../types';
 
 import type {TooltipPosition, TooltipPositionParams} from './types';
 import {DEFAULT_TOOLTIP_POSITION} from './constants';