+++ packages/polaris-viz/src/components/VerticalBarChart/utilities/get-rounded-border-for-stacked-values.ts
@@ -1,4 +1,4 @@
-import {RoundedBorder} from '../../../types';
+import {RoundedBorder} from '@shopify/polaris-viz-core';
 
 export function getRoundedBorderForStackedValues(
   values: number[],