+++ packages/polaris-viz/src/hooks/use-theme-series-colors.ts
@@ -1,12 +1,14 @@
 import {useMemo} from 'react';
-import type {Color} from '@shopify/polaris-viz-core';
+import type {
+  Theme,
+  Color,
+  DataSeries,
+  DataPoint,
+  LineStyle,
+} from '@shopify/polaris-viz-core';
 
-import type {Theme, LineStyle, DataPoint, Data} from '../types';
-
-// Note: This is a bandaid until all component are using
-// the new DataSeries type.
 interface ValidData {
-  data: (DataPoint | Data)[];
+  data: (DataPoint | DataSeries)[];
   color?: Color;
   isComparison?: boolean;
   name?: string;