+++ packages/polaris-viz/src/types.ts
@@ -1,29 +1,5 @@
 import type {InterpolatorFn} from '@react-spring/web';
-import type {Color, BarTheme} from '@shopify/polaris-viz-core';
 import type {Series, SeriesPoint} from 'd3-shape';
-
-export interface DataPoint {
-  key: number | string;
-  value: number | null;
-}
-
-export interface Data {
-  label: string;
-  rawValue: number;
-}
-
-export interface NullableData {
-  label: string;
-  rawValue: number | null;
-}
-
-export type LineStyle = 'dashed' | 'solid' | 'dotted';
-
-export interface GradientStop {
-  offset: number;
-  color: string;
-  stopOpacity?: number;
-}
 
 export type LabelFormatter = (value: string | number | null) => string;
 
@@ -35,20 +11,10 @@ export type StringLabelFormatter = (
 
 export type NumberLabelFormatter = (value: number) => string;
 
-export interface GradientStop {
-  offset: number;
-  color: string;
-}
-
 export interface YAxisTick {
   value: number;
   formattedValue: string;
   yOffset: number;
-}
-
-export interface Dimensions {
-  width: number;
-  height: number;
 }
 
 export interface SparkChartData {
@@ -77,88 +43,6 @@ export enum BarMargin {
   None = 0,
 }
 
-export interface GridTheme {
-  showHorizontalLines: boolean;
-  showVerticalLines: boolean;
-  horizontalOverflow: boolean;
-  color: string;
-  horizontalMargin: number;
-}
-
-export interface XAxisTheme {
-  showTicks: boolean;
-  labelColor: string;
-  hide: boolean;
-}
-
-export interface CrossHairTheme {
-  color: string;
-  width: number;
-}
-
-export interface YAxisTheme {
-  labelColor: string;
-  backgroundColor: string;
-}
-
-export interface ChartContainerTheme {
-  borderRadius: string;
-  padding: string;
-  backgroundColor: string;
-}
-
-export interface LineTheme {
-  sparkArea: Color | null;
-  hasSpline: boolean;
-  style: LineStyle;
-  hasPoint: boolean;
-  width: number;
-  pointStroke: string;
-}
-
-export interface TooltipTheme {
-  backgroundColor: string;
-  valueColor: string;
-  labelColor: string;
-}
-export interface SeriesColors {
-  comparison: string;
-  single: Color;
-  upToFour: Color[];
-  fromFiveToSeven: Color[];
-  all: Color[];
-}
-export interface Legend {
-  labelColor: string;
-  valueColor: string;
-  trendIndicator: {positive: string; negative: string; neutral: string};
-}
-
-export interface PartialTheme {
-  chartContainer?: Partial<ChartContainerTheme>;
-  bar?: Partial<BarTheme>;
-  line?: Partial<LineTheme>;
-  grid?: Partial<GridTheme>;
-  xAxis?: Partial<XAxisTheme>;
-  yAxis?: Partial<YAxisTheme>;
-  crossHair?: Partial<CrossHairTheme>;
-  legend?: Partial<Legend>;
-  seriesColors?: Partial<SeriesColors>;
-  tooltip?: Partial<TooltipTheme>;
-}
-
-export interface Theme {
-  chartContainer: ChartContainerTheme;
-  bar: BarTheme;
-  grid: GridTheme;
-  xAxis: XAxisTheme;
-  yAxis: YAxisTheme;
-  line: LineTheme;
-  crossHair: CrossHairTheme;
-  legend: Legend;
-  seriesColors: SeriesColors;
-  tooltip: TooltipTheme;
-}
 export interface Margin {
   Top: number;
   Left: number;
@@ -166,15 +50,6 @@ export interface Margin {
   Bottom: number;
 }
 
-export enum RoundedBorder {
-  None,
-  Top,
-  Right,
-  Bottom,
-  Left,
-}
-
-export type Direction = 'horizontal' | 'vertical';
 export type StackedSeries = Series<
   {
     [key: string]: number;