+++ packages/polaris-viz/src/utilities/change-color-opacity.ts
@@ -1,6 +1,5 @@
 import {color as d3Color} from 'd3-color';
-
-import type {GradientStop} from '../types';
+import type {GradientStop} from '@shopify/polaris-viz-core';
 
 export function changeColorOpacity(color: string, opacity = 1): string {
   const rgbColor = d3Color(color);